pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dbComm"
version = "0.0.1"
authors = [
  { name="Alex Caviness", email="caviness2@llnl.gov" },
]
description = "dbComm module for communicating with a (MongoDB) database"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  'matplotlib>=3.4.3',
  'Pillow>=9.0.1',
  'pymongo>=3.12.0',
]

[project.urls]
"Homepage" = "https://mybitbucket.llnl.gov/projects/PREC/repos/database-comm/"
